Move some code stolen from libwayland into a .c file
